requirements.txt
feedparser
supabase
python-dotenv
httpx